Don't need pickled dicts for R2 to R1 mapping
Includes slight refactoring to pass a set of identifiers to the
filtering function.

My motivation is to reduce the input files to kodoja_retrieve.py
which would facilitate adding this to Galaxy without having to
introduce any more datatypes (e.g. pickle files, or a composite
datatype for the kodoja_search.py output folder - in general
should only need the input reads and a tabular table?)

In my attempts to test this part of the code in isolation
with a real sized dataset, this is also much faster.