v0.0.8 - Output read ID not descr in kraken_VRL.txt
Doing this means the SeqID column in kraken_VRL.txt is
just the sequence ID, and then can more easily be used
downstream (e.g. in Galaxy to filter sequences by ID
as an alternative to running kodoja_retrieve.py)

Note the results are at the read pair level anyway, so
we don't need to worry about discarding the current
Illumina pair naming 1:... or 2:... in the description.

This more closely matches what Kraken does if given the
FASTQ files directly.